+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed parsing of logical lines with sequential `<` and `>` comparisons.
+- Fixed formatting of `interface` as a generic constraint.
 
 ### Added